Replace add file / add directory with one general add files and hook into uploader state.
